# api/openapi.yaml
openapi: 3.1.0
info:
  title: bazNTMS API
  version: "1.1.0"
  description: >
    bazNTMS hub HTTP API — agent filosu telemetrisi, cihaz/NetFlow/syslog
    toplama, uyarılar, RBAC ve 5651/ISO 27001 uyumluluk.


    Kimlik doğrulama açıkken (`-auth-password` veya kullanıcı DB'si) korumalı
    uçlar `nm_session` çerezi **veya** `Authorization: Bearer <token>` ister.
    Agent uçları (`/api/v1/agent/*`) UI oturumundan bağımsızdır: enrollment ile
    verilen agent token'ı ya da (hub `-tls` ile çalışıyorsa) doğrulanmış istemci
    sertifikası kullanır.


    Bu doküman elle bakımlıdır; `internal/server` içindeki bir sözleşme testi
    her yolun gerçek yönlendiricide var olduğunu doğrular. İnsan tarafı ayrıntı
    ve örnekler için `docs/API.md`.
  license:
    name: Kaynak depo
servers:
  - url: /
    description: Hub'ın kendisi (aynı origin)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/login:
    post:
      tags: [auth]
      summary: Oturum aç
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                username: { type: string, description: RBAC kullanıcısı; boşsa legacy admin }
                password: { type: string }
      responses:
        "200":
          description: "Oturum kuruldu; Set-Cookie: nm_session"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  token: { type: string }
                  role: { type: string }
                  username: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/auth/oidc/login:
    get:
      tags: [auth]
      summary: OIDC/SSO akışını başlat (sağlayıcıya yönlendirir)
      security: []
      responses:
        "302": { description: Sağlayıcıya yönlendirme }
        "404": { description: OIDC yapılandırılmamış }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nm_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Login token'ı (`token`) veya entegrasyon API token'ı (`bnt_...`) ya da agent token'ı
    mtlsAuth:
      type: mutualTLS
      description: Hub `-tls` ile çalışırken agent istemci sertifikası
...
  responses:
    Unauthorized:
      description: Oturum gerekli
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              auth_required: { type: boolean }
...
    RateLimited:
      description: Çok fazla başarısız deneme
      headers:
        Retry-After:
          schema: { type: integer }
...